move useImportTemplate into its own file
